Refactor structure, combine agent automigrate, verify (#5)
* remove automigrate and verify pkg move it to dbtool

* remove unknown name pg in sqlmapper
