+++ cmd/backend/main.go
@@ -21,7 +21,7 @@ var (
 )
 
 func main() {
-	cfg, err := backend.NewConfig(backendConfig.NewYAMLConfigReader("example_dashboard_config.yaml"))
+	cfg, err := backend.NewConfig(backendConfig.ReadYAML("example_dashboard_config.yaml"))
 	if err != nil {
 		panic(err)
 	}
